backend/pyproject.toml
[project]
name = "zhivoy-backend"
version = "0.1.0"
description = "Zhivoy backend API (FastAPI + Postgres)"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.10.4",
  "pydantic-settings==2.7.0",
  "SQLAlchemy==2.0.36",
  "psycopg[binary]==3.2.3",
  "alembic==1.14.0",
  "passlib[bcrypt]==1.7.4",
  "python-jose==3.3.0",
  "httpx==0.27.2",
]

[tool.uvicorn]
factory = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]


